Add centered and fixed highway generation
- Now the highway can generate three lanes in the same center position at all times.

- I ended up keeping a gameobject cube on the player and then attaching another gameobject cube on the 'highway' prefab I instantiate with each pass of the trigger zone.

- The cube attached to the player is to know the Z-position of where to spawn the next instance 'highway'. It moves with the player at a fixed distance away. So the generation is not dependent on WHEN the player passes through the trigger.

- However, I realized that I could not rely on the X and Y of the game object cube attached to the player. It works perfectly for the Z position, but its X and Y change with the player.

This is where the gameobject attached to the edge of 'highway' prefab comes in. By using the gameobject's X and Y positions every time a new highway generates, the highway will stay centered. I just make sure there is never more than one of that game object in the scene with an Array with all of that gameobject with that tag in it, and destroying the oldest instance of it, which would be at index 0.